Debian: Simplify /boot mounting
The original idea here was that bpool/BOOT/debian should be able to be
cloned, just like rpool/ROOT/debian. The problem is that it doesn't
really work without special support in the initramfs. Given that, it is
possible to simplify this.

I also set mountpoint=/boot on bpool, for consistency with the Ubuntu
20.04 HOWTO, which is in turn done that way for consistency with the
Ubuntu installer.

Thanks: colmbuckley
Closes: openzfs/zfs#10309
Signed-off-by: Richard Laager <rlaager@wiktel.com>