Debian/Ubuntu: Autonumber (current) HOWTOs
This also involved a bunch of reindenting and rewrapping.

I reworked the Live CD sources on Debian to be more consistent with
Ubuntu (and the chroot sources configuration on Debian).

I reworked the zpool create wrapping and the mirror/raidz notes.  This
should be more clear on how to create the multi-disk topologies.

I fixed a couple formatting issues too (mainly one backtick where
there should have been two).

Signed-off-by: Richard Laager <rlaager@wiktel.com>