Changed typo in zpool create mirror
There was a typo in the section where you create a mirror zpool. Someone accidently copied it  from the boot pool & forgot to change it. A friend and i were working on a mirror ZFS system voor a school project & we got a error that the bpool already existed, then he pointed out that there was a typo. We decided it would be a good thing to tell you guys this ;) 

Old:
       zpool create \
           ... \
           brpool mirror \
           /dev/disk/by-id/scsi-SATA_disk1-part3 \
           /dev/disk/by-id/scsi-SATA_disk2-part3

Should be:
       zpool create \
           ... \
           rpool mirror \
           /dev/disk/by-id/scsi-SATA_disk1-part4 \
           /dev/disk/by-id/scsi-SATA_disk2-part4