Ubuntu: Remove canmount=noauto on / and /boot
I think (but am not absolutely certain) that things behave as follows:

- The non-zsys initramfs script expects canmount=noauto for the root
  filesystem.  It then mounts the active one (e.g. in the case of
  clones) manually, overriding that.
- zsys manages the canmount attribute.

I am sure that when the system boots with zsys, the initial datasets,
which were created with canmount=noauto, have canmount=on.

Therefore, there seems to be no reason to set canmount=noauto for the
zsys scenario, and I have removed it.  This simplifies the instructions
and may avoid issues like #73.

Signed-off-by: Richard Laager <rlaager@wiktel.com>
Fixes #73