Ubuntu: Go back to bpool/grub
For a mirror or raidz topology, /boot/grub is on a separate dataset.
This was originally bpool/grub, then changed on 2020-05-30 to
bpool/BOOT/ubuntu_UUID/grub to work-around zsys setting canmount=off
which would result in /boot/grub not mounting.  This work-around lead
to issues with snapshot restores. The underlying zsys issue
ubuntu/zsys#164 was fixed and backported to 20.04, so it is now back to
being bpool/grub.

This reverts commit b6fd009eddcc6f97d7627e5727f7d62b7f605d14.  It
conceptually reverts the errata notes from commit
04d3c1cee442ea964248029a2b292fc5ebed0bdd, but includes new steps for
people in the pre-2020-05-30 state.

Signed-off-by: Richard Laager <rlaager@wiktel.com>
Fixes #55