compatibility maxtrix: Add Nexenta
This might be useful to someone.  It definitely would have been useful
for me in the past.  I have a Nexenta 4 system in production, so I was
able to personally verify the feature support there.

This has a couple of hacks, like with other sources, to workaround
documentation bugs.

Fixes #82

Signed-off-by: Richard Laager <rlaager@wiktel.com>