Debian/Ubuntu: Try again with /run
The fix for #97, 66233a11389d392094d7a17c97a9a88dbc775cc8, seems to
have created yet another issue (beyond #100):

"the link /etc/resolv.conf -> ../run/systemd/resolve/stub-resolv.conf
is dangling (since /run has just been clobbered by an empty tmpfs).
debootstrap seems to populate all manner of stuff in /run."
https://github.com/openzfs/openzfs-docs/commit/f1721edda7ccca6fdcf397086b88c71f292073a8#commitcomment-45770519

I considered just reverting all of this, since the original issue only
occurred when copying in an existing system.  Ignoring it was my
decision the first time this came up.  But it's come up twice now.
Also, this mode is mentioned in the guide.  Most importantly, if we are
not mounting a /tmpfs at /run in the normal case, then we could end up
with cruft in /run in the root filesystem which could potentially break
something or be confusing later if the /run tmpfs ever failed to mount.
So in the interest of correctness, I'd like to try to handle this.

I am now mounting a tmpfs at /mnt/run before the debootstrap step.

Signed-off-by: Richard Laager <rlaager@wiktel.com>