Debian/Ubuntu: Update /boot/efi fstab entry
I made Debian follow the Ubuntu approach, which comes from the Ubuntu
installer.  There was a report (#99) of a problem with the old approach
used on Debian.  In that issue, user foo-git linked to a Fedora feature
request which discusses automounting /boot/efi instead.  I've offered
that as an alternative.

I am not switching to that outright, as following the Ubuntu installer
(where possible) is a goal of mine.

Suggested-by: foo-git
Closes #99