Fix ZFS feature generation script
This was breaking on FreeBSD, as some of the URLs were returning 404.  That
raises an HTTPError.  If we catch those and skip those sources (just as we
were already doing for HTTP codes other than 200), the script completes.  But
then we get a whole bunch of columns full of "no", which does not seem useful.
I added code to skip columns for sources that failed to download.

Signed-off-by: Richard Laager <rlaager@wiktel.com>