Debian: Fix /run/lock code
It looks like I messed this up (for Debian, but not Ubuntu) in my
various changes.  It is not necessary to do this before the chroot in
the normal install path, as it was done earlier.  But it is something
that should be done before the chroot in the rescue path.  I've now
made Debian match Ubuntu.

Fixes #141
Reported-by: Colm <colm@tuatha.org>
Signed-off-by: Richard Laager <rlaager@wiktel.com>