Add more packages to the APT preference file
When doing `apt upgrade` on my Debian machine, I saw that two packages was being held back. The reason cited was that the names had somewhat changed I think. It is worth noting that I'm running bullseye, so this might not be applicable to buster, which is the version documented here.