do *not* enforce UTF-8
I have tried to convert a real, live workstation to ZFS and had to
recreate all filesystems because of this silly settings. It actually
failed to copy [source code from the supysonic project][1] (which was
[eventually removed][2], but still).

This commit addresses *all* such incantations I could find, but keep a
reference to it in the **Hints** section or, if it's not present,
the **Notes** section, mentioning it was removed and why. This is so
people *can* still add it if they want.

 [1]: https://github.com/spl0k/supysonic/tree/270fa9883b2f2bc98f1482a68f7d9022017af50b/tests/assets/%E
 [2]: https://github.com/spl0k/supysonic/pull/183