Debian & Ubuntu: Rework dataset creation step
The major change for Ubuntu is to mark many of these as optional, like
with Debian and the old guides.  As I move away from zsys, this will end
up being more like the old way.

This was also discussed in #306 and #307, but this change has
trade-offs.  It can reduce the number of datasets created on the
system, but it does so by increasing the complexity to read and follow
the guide.

Then I just harmonized Debian with Ubuntu.  Aside from whitespace,
reordering, rewording, the substantive changes were to drop /opt and
add /var/lib/NetworkManager.

Signed-off-by: Richard Laager <rlaager@wiktel.com>