reorder and regroup zpool create commands
This regroups the options passed to zpool create in a more logical
way.

 1. the ACL/xattr/dnodesize settings are all on one line, since they
    are all related (latter being optimizations because we enable the
    former)

 2. ashift, autotrim, compression, normalization, relatime are all
    independent options that merit their own line for better
    visibility, they are also optional and therefore it makes sense to
    have them separate

 3. canmount, mountpoint and -R are all related as well

This is only a "whitespace" fix: there should be no functional change
to this patch.

I also noted that dnodesize=auto is *not* passed to the bpool cache
creation, even though it's passed to the rpool creation. I suspect
that might be an omission. This also goes to show that ordering those
options consistently makes it easier to spot those errors.

This is a followup to #307.

Signed-off-by: Antoine Beaupré <anarcat@debian.org>