getting started: opensuse: mention zfs-ueficert package
openSUSE enables kernel_lockdown(7), which forces all kernel modules to
need to be signed. The kernel module package building infrastructure
can sign all modules (and does for ZFS), but the module is not signed
with the official project keys (as it is managed by a devel project).

As per [1] the solution is to simply install zfs-ueficert which contains
the right signing key and also will auto-register the key with MOK on
the next boot.

[1]: https://bugzilla.suse.com/show_bug.cgi?id=1173551

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>